Neural Search Engine/requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
fastapi
uvicorn
streamlit
requests